refactor: split JSON-LD schema construction out of seoUtils
seoUtils.ts (780 lines) mixed two unrelated concerns: schema.org
JSON-LD schema types/generators and SEO description/keyword generation.
Move the JSON-LD half verbatim into jsonLdSchemas.ts with its own
imports. seoUtils.ts keeps the SEO-text half and re-exports
jsonLdSchemas (export *), so all 14 existing `$lib/utils/seoUtils`
consumers keep working with zero changes and zero behavior change.
